feat: move `Point` to be usable by other algorithms and correct the way Nelder-Mead functions are evaluated with bounds
